fix: resolve max-lines lint errors in scheduled-event and settings services
Extract resolveVoiceForEdit to discord-ops helpers, remove re-exports
from settings.service.ts and update consumers to import from types directly.
