+++ api/src/discord-bot/services/scheduled-event.discord-ops.ts
@@ -4,8 +4,16 @@ import {
   timedDiscordCall,
 } from './scheduled-event.helpers';
 import type { DiscordBotClientService } from '../discord-bot-client.service';
+import type { ScheduledEventRecord } from './scheduled-event.db-helpers';
 
 type Guild = NonNullable<ReturnType<DiscordBotClientService['getGuild']>>;
+
+interface VoiceChannelResolver {
+  resolveVoiceChannelForScheduledEvent(
+    gameId?: number | null,
+    recurrenceGroupId?: string | null,
+  ): Promise<string | null>;
+}
 
 export function isUnknownEventError(error: unknown): boolean {
   return (
@@ -150,4 +158,27 @@ export async function tryEditDescription(
     if (isUnknownEventError(error)) return true;
     throw error;
   }
+}
+
+/**
+ * Resolve the voice channel for a scheduled event edit (ROK-716).
+ * If notificationChannelOverride is set and is a voice channel, use it.
+ * If it's a text channel, fall back to the channel resolver.
+ * If it's not in cache, use it optimistically (may be an uncached voice channel).
+ */
+export async function resolveVoiceForEdit(
+  guild: Guild,
+  event: ScheduledEventRecord,
+  gameId: number | null | undefined,
+  channelResolver: VoiceChannelResolver,
+): Promise<string | null> {
+  const override = event.notificationChannelOverride;
+  if (override) {
+    const cached = guild.channels.cache.get(override);
+    if (!cached || cached.isVoiceBased()) return override;
+  }
+  return channelResolver.resolveVoiceChannelForScheduledEvent(
+    gameId,
+    event.recurrenceGroupId,
+  );
 }
