+++ api/src/discord-bot/services/scheduled-event.service.ts
@@ -33,6 +33,7 @@ import {
   tryCompleteEvent,
   tryEditEndTime,
   tryEditDescription,
+  resolveVoiceForEdit,
 } from './scheduled-event.discord-ops';
 
 export type { ScheduledEventData } from './scheduled-event.helpers';
@@ -278,7 +279,12 @@ export class ScheduledEventService {
     gameId?: number | null,
   ): Promise<void> {
     const description = await this.buildDescription(eventId, eventData);
-    const voiceChannelId = await this.resolveVoiceForEdit(guild, event, gameId);
+    const voiceChannelId = await resolveVoiceForEdit(
+      guild,
+      event,
+      gameId,
+      this.channelResolver,
+    );
     try {
       await timedDiscordCall(
         'scheduledEvents.edit',
@@ -318,28 +324,6 @@ export class ScheduledEventService {
         recurrenceGroupId,
       )) ??
       null
-    );
-  }
-
-  /**
-   * Resolve the voice channel for a scheduled event edit (ROK-716).
-   * If notificationChannelOverride is set and is a voice channel, use it.
-   * If it's a text channel, fall back to the channel resolver.
-   * If it's not in cache, use it optimistically (may be an uncached voice channel).
-   */
-  private async resolveVoiceForEdit(
-    guild: NonNullable<ReturnType<DiscordBotClientService['getGuild']>>,
-    event: ScheduledEventRecord,
-    gameId?: number | null,
-  ): Promise<string | null> {
-    const override = event.notificationChannelOverride;
-    if (override) {
-      const cached = guild.channels.cache.get(override);
-      if (!cached || cached.isVoiceBased()) return override;
-    }
-    return this.channelResolver.resolveVoiceChannelForScheduledEvent(
-      gameId,
-      event.recurrenceGroupId,
     );
   }
