+++ api/src/igdb/igdb.service.ts
@@ -12,7 +12,8 @@ import * as schema from '../drizzle/schema';
 import { stripSearchPunctuation } from '../common/search.util';
 import type { IgdbGameDto, GameDetailDto } from '@raid-ledger/contract';
 import type { ActivityPeriod } from '@raid-ledger/contract';
-import { SettingsService, SETTINGS_EVENTS } from '../settings/settings.service';
+import { SettingsService } from '../settings/settings.service';
+import { SETTINGS_EVENTS } from '../settings/settings.types';
 import { SETTING_KEYS } from '../drizzle/schema/app-settings';
 import { IGDB_SYNC_QUEUE, IgdbSyncJobData } from './igdb-sync.constants';
 import { CronJobService } from '../cron-jobs/cron-job.service';
